requirements.txt
# Conda packages (Miguel)
numpy
pandas
PyYAML

# Pip packages (Miguel)
Pyomo

